+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## What's Changed
 
+- Add install mode virtual data point with countdown timer (HUB_BUTTON + HUB_SENSOR)
 - Add get_install_mode/set_install_mode to DeviceManagement/CentralUnit
 
 # Version 2025.12.4 (2025-12-02)